Fix initializing of data grid
When opening in a docked-out window, it could happen that the data binding for the ItemsSource was evaluated after the Loaded event of the DataGrid was fired. In that case the collection wasn't yet initialized and we got a NullReferenceException while trying to apply the layout to the collection.

Additionally, when ItemsSource was changed (which currently never occurs), no layout was applied to new newly created collection.